% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate your machine with Wrike identifiers}
\usage{
authenticate()
}
\description{
Loads Wrike authenticate keys from Sys.getenv as account_id and v3_key
}
\examples{
authenticate()

}
