% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{get_wrike_comments}
\alias{get_wrike_comments}
\title{Get all comments in the past 7 days}
\usage{
get_wrike_comments()
}
\description{
Get dataframe of comments associated with your Wrike account
}
\examples{
get_wrike_comments()
}
\keyword{}
