% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{get_wrike_task_comment}
\alias{get_wrike_task_comment}
\title{Get comments on Wrike task}
\usage{
get_wrike_task_comment(task_id)
}
\arguments{
\item{task_id}{Id associated with the task you want to create a comment on (ex:IEABOGRQKQAN3QOA)}
}
\description{
Get dataframe of comments on a specified Wrike task
}
\examples{
get_wrike_task_comment("My Task Id")
get_wrike_task_comment("IEABOGRQKQAN3QOA")
}
