% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{wrike_account_data}
\alias{wrike_account_data}
\title{Wrike Account Information}
\usage{
wrike_account_data()
}
\value{
A tibble with the following account fields: id, name, dateFormat, rootFolderId, createdDate
}
\description{
This function queries Wrike account information.
}
\examples{
wrike_account_data()
}
