% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{wrike_account_id}
\alias{wrike_account_id}
\title{Wrike Account ID}
\usage{
wrike_account_id()
}
\description{
This function pulls the organization's Wrike account ID for authentication.
}
\examples{
wrike_account_id()
}
