% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{wrike_custom_field_on_task}
\alias{wrike_custom_field_on_task}
\title{Wrike Custom Field Use by Task Id}
\usage{
wrike_custom_field_on_task(task_id, custom_field_id)
}
\arguments{
\item{task_id}{Wrike task id}

\item{custom_field_id}{Use \code{\link{wrike_custom_field_url}} function to find id if needed.}
}
\description{
This function pulls a list of the custom field values & ids associated with specified task ids
}
\examples{
wrike_custom_field_on_task(task_id = "IEABOGRQKQAN3QOA", custom_field_id = "IEAAAOH5JUAAABQ5")

}
