% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{wrike_custom_field_update}
\alias{wrike_custom_field_update}
\title{Wrike Custom Field Update}
\usage{
wrike_custom_field_update(task_id, custom_field_id, custom_field_value)
}
\arguments{
\item{task_id}{Wrike task id}

\item{custom_field_id}{Use \code{\link{wrike_custom_field_url}} function to find id if needed.}

\item{custom_field_value}{What you want populated}
}
\description{
This function populates a custom field from specified task id
}
\examples{
wrike_custom_field_update(task_id = "IEABOGRQKQAN3QOA", custom_field_id = "IEAAAOH5JUAAABQ5", custom_field_value = "myvalue") <-- UPDATE for gh

}
