% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_fields.R
\name{wrike_custom_field_url}
\alias{wrike_custom_field_url}
\title{Wrike Custom Field URL}
\usage{
wrike_custom_field_url()
}
\description{
This function gives you the URL to access your account's custom fields
}
\examples{
wrike_custom_field_url()
DEPRECATED
}
