% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folders.R
\name{wrike_folder_id}
\alias{wrike_folder_id}
\title{Wrike Folders IDs}
\usage{
wrike_folder_id(folder_name)
}
\description{
This function grabs the folder id when given the folder name.
}
\examples{
wrike_folder_id("My Folder Name")

}
