% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folders.R
\name{wrike_folder_tree}
\alias{wrike_folder_tree}
\title{Wrike Folder Tree}
\usage{
wrike_folder_tree(folder_name)
}
\arguments{
\item{folder_name}{Name of root folder to pull child ids from}
}
\description{
Pulls list of all child folders and folder ids from parent folder identified
}
\examples{
wrike_folder_tree("My Folder Name")
}
