% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_folders.R
\name{wrike_folders}
\alias{wrike_folders}
\title{Wrike Folders}
\usage{
wrike_folders()
}
\description{
This function pulls a list of your Wrike folder names + ids. Ids will be used in task functions
}
\examples{
wrike_folders()

}
