% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{wrike_get_comments}
\alias{wrike_get_comments}
\title{Get Wrike Comments from folder}
\usage{
wrike_get_comments(folder_id)
}
\arguments{
\item{folder_id}{Wrike folder id that you want to get comments from. Find ids using 'wrike_folders()'}
}
\description{
Pulls a list of comments associated with specified task. Last 7 days only.
}
\examples{
wrike_get_comments("My Folder ID")
}
