% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tasks.R
\name{wrike_task_data}
\alias{wrike_task_data}
\title{Wrike Tasks by ID}
\usage{
wrike_task_data(task_id, fields = NULL)
}
\arguments{
\item{task_id}{Wrike task id}

\item{fields}{Defaults to all, you can select a specific field by name}
}
\description{
Pulls a list of Wrike task information by the Wrike task id
}
\examples{
wrike_task_data("IEABOGRQKQAN3QOA")
wrike_task_data("IEABOGRQKQAN3QOA", fields = "title")
wrike_task_data("IEABOGRQKQAN3QOA", fields = c("title", "status", "id"))

}
