% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tasks.R
\name{wrike_tasks}
\alias{wrike_tasks}
\title{Wrike Tasks}
\usage{
wrike_tasks(folder_name)
}
\arguments{
\item{folder_name}{}
}
\description{
Pulls a list of all Wrike tasks based on specified folder
}
\examples{
wrike_tasks("My Folder Name")
}
