% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comments.R
\name{wrike_url_comment_post}
\alias{wrike_url_comment_post}
\title{Create simple comment on Wrike task}
\usage{
wrike_url_comment_post(taskId, commentText)
}
\arguments{
\item{taskId}{Id associated with the task you want to create a comment on (ex:IEABOGRQKQAN3QOA)}

\item{commentText}{short, single word comment}
}
\description{
Pulls a list of comments associated with specified task. POST function allows for comment creation on specified task
}
\examples{
wrike_url_comment_post("My Task ID", "hello")
wrike_url_comment_post("IEABOGRQKQAN3QOA", "hello")
}
\keyword{}
