% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_workflows.R
\name{wrike_workflows}
\alias{wrike_workflows}
\title{Wrike Workflows}
\usage{
wrike_workflows()
}
\description{
Pulls a list of all Wrike workflow labels + adds a progress tracker
}
\examples{
wrike_workflows()
}
