% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrikeR-package.R
\docType{package}
\name{wriker-package}
\alias{wriker}
\alias{wriker-package}
\title{wriker: Wrike Project Management API Wrapper for R}
\description{
An API wrapper for pulling data from Wrike's web API via R. Wrike documentation here
    (<https://developers.wrike.com/>). Package allows enterprise users to
    retreive information on their task, folder, comment, workflow, and custom 
    field data.
}
\author{
\strong{Maintainer}: Alexandra Mangold \email{alexandra.a.mangold@gmail.com}

}
\keyword{internal}
